requirements.txt
playwright
playwright-stealth
openai
python-dotenv
